Updated version of dependencies. Updated version of gradle.
