+++ app/src/main/java/com/makedonsky/example/ViewPagerActivity.java
@@ -1,13 +1,18 @@
 package com.makedonsky.example;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
 
 import com.makedonsky.widget.SwipeToClose;
 
@@ -51,5 +56,41 @@ public class ViewPagerActivity extends AppCompatActivity {
                 .with(this)
                 .withSensitivity(0.1f)
                 .bind();
+    }
+
+    private class ColorPagerAdapter extends PagerAdapter {
+        int[] colors = {
+                R.color.colorAccent,
+                R.color.colorPrimary,
+                R.color.colorPrimaryDark
+        };
+
+        @Override
+        public int getCount() {
+            return colors.length;
+        }
+
+        @Override
+        public boolean isViewFromObject(View view, Object object) {
+            return view == object;
+        }
+
+        @Override
+        public Object instantiateItem(ViewGroup container, int position) {
+            Context context = container.getContext();
+            ImageView itemView = (ImageView) LayoutInflater
+                    .from(context)
+                    .inflate(R.layout.image, container, false);
+            itemView.setImageResource(colors[position]);
+
+            container.addView(itemView);
+
+            return itemView;
+        }
+
+        @Override
+        public void destroyItem(ViewGroup container, int position, Object object) {
+            container.removeView((View) object);
+        }
     }
 }
