+++ swipe-to-close/build.gradle
@@ -1,18 +1,37 @@
 apply plugin: 'com.android.library'
 
 ext {
-    PUBLISH_GROUP_ID = 'com.makedonsky.widget'
-    PUBLISH_ARTIFACT_ID = 'swipe-to-close'
-    PUBLISH_VERSION = '0.3.9'
+    bintrayRepo = 'widget'
+    bintrayName = 'swipe-to-close'
+
+    publishedGroupId = 'com.makedonsky.widget'
+    libraryName = 'SwipeToClose'
+    artifact = 'swipe-to-close'
+
+    libraryDescription = 'SwipeToClose is the library which using for create an Activity with swipe-to-close functionality.'
+
+    siteUrl = 'https://github.com/makedonsky94/SwipeToClose'
+    gitUrl = 'https://github.com/makedonsky94/SwipeToClose.git'
+
+    libraryVersion = '0.3.9'
+
+    developerId = 'makedonsky'
+    developerName = 'Alexandr Nevyantsev'
+    developerEmail = 'nevyantsev.alexandr@gmail.com'
+
+    licenseName = 'The Apache Software License, Version 2.0'
+    licenseUrl = 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+    allLicenses = ["Apache-2.0"]
 }
 
+
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.3"
+    compileSdkVersion 25
+    buildToolsVersion "25.0.3"
 
     defaultConfig {
         minSdkVersion 17
-        targetSdkVersion 24
+        targetSdkVersion 25
         versionCode 1
         versionName "0.3.9"
     }
@@ -22,13 +41,13 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-
-    return void
 }
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:24.2.1'
+    compile 'com.android.support:appcompat-v7:25.3.1'
 }
 
-apply from: 'https://raw.githubusercontent.com/blundell/release-android-library/master/android-release-aar.gradle'
+
+apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
+apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
